Run on Windows on Python 3.5

+++ ast27/Include/asdl.h
@@ -34,8 +34,8 @@ typedef struct {
 #define _Py_asdl_seq_new asdl_seq_new
 #define _Py_asdl_int_seq_new asdl_int_seq_new
 #endif
-asdl_seq *asdl_seq_new(int size, PyArena *arena);
-asdl_int_seq *asdl_int_seq_new(int size, PyArena *arena);
+asdl_seq *asdl_seq_new(Py_ssize_t size, PyArena *arena);
+asdl_int_seq *asdl_int_seq_new(Py_ssize_t size, PyArena *arena);
 
 #define asdl_seq_GET(S, I) (S)->elements[(I)]
 #define asdl_seq_LEN(S) ((S) == NULL ? 0 : (S)->size)
